Nonlinear convergence archived by tightening tolerance of linear system
When eliminate pressure values, these are moved to the right-hand side of the
system.

This means that the norm of the right-hand side changes with many orders of
magnitude compared to what it typically is.  As a result, the tolerances in the
linear solver (BiCGSTAB) needs to be tighened a lot.