Use Well Fracturing Seeds From SCHEDULE Section
This commit switches the "well_seed" fracture initialisation method
to getting the seed point and associate fracturing plane normal
vectors from the WSEED keyword in the SCHEDULE section.  At this
time, there is an implicit assumption that the WSEED keyword is
entered at the beginning of the SCHEDULE section, but we will relax
this assumption in follow-up work.

As an aid to future development, we split each fracture
initialisation method out to a separate member function and call the
pertinent member function depending on the PropertyTree's
"fractureparam.config.type", with the "well_seed" method being the
default.

Note that as the WSEED keyword is only available in 2025.10 and
later, we also switch the minimum compatibility level to 2025.10-pre
as part of this work.