Tag Extra Connections as Dynamic Fractures
This ensures that these extra connections will not be output to the
restart file and therefore not create confusing displays in
post-processing tools like ResInsight or adversely affect simulation
restart.

While here, ensure that we don't call applySimulatorUpdate() if
there is no update to apply--i.e., if there are no extra
connections.