Generate MSW Segment Information for Runtime Connections
This commit introduces support for providing minimal connection
level segment information--segment numbers and tubing lengths--for
well connections generated at runtime, typically from a fracturing
process.

To this end, introduce a new helper type

    FractureWell::Connection

containing the reservoir cell, the associate segment number if
applicable (zero if not an MSW), and the tubing lengths (nullopt if
not an MSW).  Construct the FractureWell object from a vector of
these connections instead of a vector of reservoir cells and provide
FractureWell accessors for the additional data members.  Then, in
EclProblemGeoMech<>::addConnectionsToSchedual(), include this
segment information in the new connection objects if applicable.