Introduce Basic Regression Testing Framework
This commit adds initial support for running regression tests.  We
split running the model and comparing the simulation results into
two separate steps in which the second depends on the first through
"FIXTURES_REQUIRED".  This initial work is mostly proof of concept
and does not actually compare the results to a known good state.
Adding that will be follow-up work, once we have tooling in place to
create and maintain reference solutions.